package platform backends for x11 and wayland and recommend qtwayland5
talked with mgraesslin about this... at install time we cannot know
whether an application will need the wayland or x11 plugin (let alone
third party applications outside of a packaging context), so the way to
ensure things work correctly is to have both x11 and wayland installed
at the same time.
this has to stay this way for the foreseeable future (possibly until
upstream decides to drop the x11 plugin) as this is sort of required to
keep kwindowsystem working for x11 apps and wayland apps. this does not
have anything to do with which display technology is actually used but
rather what the application was built around.

packaging both forcefully has the disadvantage of pulling in xcb/libwayland
on systems where one of the two would technically not be needed, which
is a bit meh but better than having broken functionality.

upstream *explicitly* requested both to be present all the time
to ensure that all applications will work correctly.