recommend kwayland-integration
while kwayland-integration is a plasma package it provides plugins needed
by frameworks to properly integrate into wayland. the plugins are moving
to frameworks (or such is the plan) when kwayland is ready to move as well.
until then recommends is the way to go since we always want this in case
an application using this framework is run inside a wayland session.